Change for (unsigned int) loops to use size_t
This makes sure that loop index is known to not overflow a 32-bit value
on a 64-bit system, and allows the compiler to produce better code.

+++ src/simplifier.cpp
@@ -79,7 +79,7 @@ static bool findEdge(const EdgeAdjacency& adjacency, unsigned int a, unsigned in
 	unsigned int count = adjacency.counts[a];
 	const unsigned int* data = adjacency.data.data + adjacency.offsets[a];
 
-	for (unsigned int i = 0; i < count; ++i)
+	for (size_t i = 0; i < count; ++i)
 		if (data[i] == b)
 			return true;
